Fixed libusb open method.

+++ qbaseusb.h
@@ -37,7 +37,7 @@ protected slots:
     void setDefaults(void);
 
 protected:
-    ulong mTimeout;
+    quint16 mTimeout;
     quint16 mPid;
     quint16 mVid;
     QString mGuid;
